Lection 1, Task1, Completed
