+++ Белов/1/Task1.java
@@ -1,0 +1,32 @@
+import java.util.Random;
+
+/*
+Задание 1. Hello, Random!
+		Завести две переменных типа Integer
+		В каждую переменную записать случайное число от 0 до 20, используя класс Random и метод nextInt(Integer bound)
+		Вывести на экран оба числа
+		Если первое число больше 10 - то вывести на экран надпись "Первое число больше десяти"
+		Если второе число четное - то вывести на экран надпись "Второе число четное". Иначе - вывести на экран надпись "Второе число нечетное"
+*/
+class Task1{
+	public static void main(String[] args){
+		Random random = new Random();
+		Integer i1 = random.nextInt(20);
+		Integer i2 = random.nextInt(20);
+		p("INtegers: " + i1 + ", " + i2); //Вывести на экран оба числа
+		if(i1>10){ //Если первое число больше 10
+			p("Первое число больше десяти"); //то вывести на экран надпись "Первое число больше десяти"
+		}
+		//Если второе число четное - то вывести на экран надпись "Второе число четное".
+		// Иначе - вывести на экран надпись "Второе число нечетное"
+		p("Второе число " + (0==i2%2 ? "" : "не") + "четное");
+	}
+	private static void p(String s){
+		System.out.println(s);
+	}
+}
+/* output:
+INtegers: 19, 3
+Первое число больше десяти
+Второе число нечетное
+ */
