--- Белов/1/task1.java
@@ -1,12 +1,0 @@
-/*
-Задание 1. Hello, Random!
-		Завести две переменных типа Integer
-		В каждую переменную записать случайное число от 0 до 20, используя класс Random и метод nextInt(Integer bound)
-		Вывести на экран оба числа
-		Если первое число больше 10 - то вывести на экран надпись "Первое число больше десяти"
-		Если второе число четное - то вывести на экран надпись "Второе число четное". Иначе - вывести на экран надпись "Второе число нечетное"
-*/
-class Task1{
-	public static void main(String[] args){
-	}
-}
